backend/requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
greenlet>=3.0.0                # required by SQLAlchemy async — enables cooperative multitasking
                               # between sync SQLAlchemy internals and the async event loop

# Redis
redis>=5.0.0

# Config
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# LLM
anthropic>=0.25.0

# RAG
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
numpy>=1.26.0                  # FAISS returns numpy arrays; sentence-transformers depends on it
torch>=2.2.0                   # sentence-transformers backend — pinned so pip picks CPU wheel, not CUDA

# Utilities
httpx>=0.27.0
tenacity>=8.3.0
python-json-logger>=2.0.0      # formats log records as JSON lines for Datadog/CloudWatch

# Testing
pytest>=8.0.0                  # test runner
pytest-asyncio>=0.23.0         # makes async test functions work with pytest
pytest-mock>=3.14.0            # cleaner mocking via the `mocker` fixture
